[FIX] website_slides: remove no more useful old activity overrides
task-2429670
