+++ addons/website_slides/static/src/components/activity/activity.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<templates xml:space="preserve">
+    <t t-inherit="mail.Activity" t-inherit-mode="extension">
+        <xpath expr="//*[@name='tools']" position="replace">
+            <t t-if="activity.requestingPartner and activity.thread.model === 'slide.channel'">
+                <div class="o_Activity_tools">
+                    <button class="o_Activity_toolButton o_Activity_grantAccessButton btn btn-link" t-on-click="_onGrantAccess">
+                        <i class="fa fa-check"/> Grant Access
+                    </button>
+                    <button class="o_Activity_toolButton o_Activity_refuseAccessButton btn btn-link" t-on-click="_onRefuseAccess">
+                        <i class="fa fa-times"/> Refuse Access
+                    </button>
+                </div>
+            </t>
+            <t t-else="">$0</t>
+        </xpath>
+    </t>
+</templates>
