--- addons/website_slides/static/src/xml/activity.xml
@@ -1,38 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<templates xml:space="preserve">
-    <t t-extend="mail.activity_items">
-        <t t-jquery=".o_thread_message .o_thread_message_core .o_thread_message_tools" t-operation="replace">
-            <t t-if="activity.activity_type_id[1] == 'Access Request'">
-                <t t-if="activity.user_id[0] === uid">
-                    <a role="button" class="btn btn-link btn-success text-muted text-success o_activity_link o_activity_action_grant_access" t-att-data-partner-id="activity.request_partner_id[0]">
-                        <i class="fa fa-check"/> Grant Access
-                    </a>
-                    <a role="button" class="btn btn-link btn-danger text-muted text-danger o_activity_link o_activity_action_refuse_access" t-att-data-partner-id="activity.request_partner_id[0]">
-                        <i class="fa fa-times"/> Refuse Access
-                    </a>
-                </t>
-            </t>
-            <t t-else="">
-                <div class="o_thread_message_tools btn-group">
-                    <t t-call="mail.activity_thread_message_tools"/>
-                </div>
-            </t>
-        </t>
-    </t>
-
-    <t t-inherit="mail.Activity" t-inherit-mode="extension">
-        <xpath expr="//*[@name='tools']" position="replace">
-            <t t-if="activity.requestingPartner and activity.thread.model === 'slide.channel'">
-                <div class="o_Activity_tools">
-                    <button class="o_Activity_toolButton o_Activity_grantAccessButton btn btn-link" t-on-click="_onGrantAccess">
-                        <i class="fa fa-check"/> Grant Access
-                    </button>
-                    <button class="o_Activity_toolButton o_Activity_refuseAccessButton btn btn-link" t-on-click="_onRefuseAccess">
-                        <i class="fa fa-times"/> Refuse Access
-                    </button>
-                </div>
-            </t>
-            <t t-else="">$0</t>
-        </xpath>
-    </t>
-</templates>
