+++ addons/website_slides/views/assets.xml
@@ -5,9 +5,9 @@
             <xpath expr="." position="inside">
                 <link rel="stylesheet" type="text/scss" href="/website_slides/static/src/scss/rating_rating_views.scss"/>
                 <link rel="stylesheet" type="text/scss" href="/website_slides/static/src/scss/slide_views.scss"/>
+                <script type="text/javascript" src="/website_slides/static/src/components/activity/activity.js"/>
                 <script type="text/javascript" src="/website_slides/static/src/js/slide_category_one2many.js"/>
                 <script type="text/javascript" src="/website_slides/static/src/js/rating_field_backend.js"/>
-                <script type="text/javascript" src="/website_slides/static/src/js/activity.js"/>
             </xpath>
         </template>
